refactor(automations): extract EventTriggerForm into its own file
Moves the ~180-line EventTriggerForm component out of automation-detail.tsx
(912 lines) into its own file alongside its sibling automation components
(add-starter-popover, trigger-card, webhook-secret-dialog). It only reads
its own props (automationId, onDone) and shares no state/closures with
SettingsTab, so this is a byte-identical relocation, not a rewrite.
